+++ src/detectors/builtin/argCopyMutation.ts
@@ -3,6 +3,7 @@ import {
   foldStatements,
   collectMutations,
   mutationNames,
+  funName,
 } from "../../internals/tact";
 import { intersection } from "../../internals/util";
 import { MistiTactWarning, Severity } from "../../internals/warnings";
@@ -16,7 +17,7 @@ import {
 
 /**
  * A detector that highlights cases where function argument mutations are ineffective
- * due to call-by-value semantics in Tact.
+ * due to [call-by-value semantics](https://en.wikipedia.org/wiki/Evaluation_strategy#Call_by_value) in Tact.
  *
  * ## Why is it bad?
  * In Tact, function arguments are passed by value, meaning that any mutations applied
@@ -54,15 +55,50 @@ export class ArgCopyMutation extends ASTDetector {
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return Array.from(cu.ast.getFunctions()).reduce((acc, fun) => {
       if (fun.kind === "contract_init" || fun.kind === "function_def") {
-        return acc.concat(
-          foldStatements(fun, [] as MistiTactWarning[], (acc, stmt) => {
+        const mutations = foldStatements(
+          fun,
+          new Map<string, AstStatement[]>(),
+          (acc, stmt) => {
             const interestingArgs = this.collectInterestingArgs(fun);
             if (interestingArgs.length === 0) {
               return acc;
             }
-            return this.findArgCopyMutations(acc, stmt, interestingArgs);
-          }),
+            const stmtMutations = this.findArgCopyMutations(
+              stmt,
+              interestingArgs,
+            );
+            return Array.from(stmtMutations.entries()).reduce(
+              (newAcc, [argName, mutationStmts]) => {
+                const existingStmts = newAcc.get(argName) || [];
+                return newAcc.set(argName, [
+                  ...existingStmts,
+                  ...mutationStmts,
+                ]);
+              },
+              new Map(acc),
+            );
+          },
         );
+        Array.from(mutations.keys()).forEach((argName) => {
+          const argMutationStatements = mutations.get(argName)!;
+          const occurrencesStr =
+            argMutationStatements.length > 1
+              ? ` (${argMutationStatements.length} more times)`
+              : "";
+          acc.push(
+            this.makeWarning(
+              `Function ${funName(fun)} argument ${argName} is mutated${occurrencesStr}`,
+              argMutationStatements[0].loc,
+              {
+                extraDescription:
+                  "Mutating function arguments has no effect outside the function due to call-by-value semantics",
+                suggestion:
+                  "Return the modified value or use the contract's state to avoid unnecessary mutations",
+              },
+            ),
+          );
+        });
+        return acc.concat();
       }
       return acc;
     }, [] as MistiTactWarning[]);
@@ -87,30 +123,22 @@ export class ArgCopyMutation extends ASTDetector {
     }, [] as string[]);
   }
 
+  /**
+   * Identifies mutations of function arguments within a given statement.
+   * @param argNames Names of function arguments to check for mutations.
+   * @returns A map of argument names to the statements where they are mutated.
+   */
   private findArgCopyMutations(
-    acc: MistiTactWarning[],
     stmt: AstStatement,
     argNames: string[],
-  ): MistiTactWarning[] {
+  ): Map<string, AstStatement[]> {
     const mutations = collectMutations(stmt);
     const foundMutations = mutations
       ? intersection(argNames, mutationNames(mutations.mutatedLocals))
       : [];
-    if (foundMutations.length > 0) {
-      const mut = foundMutations.join(", ");
-      acc.push(
-        this.makeWarning(
-          `Function argument${foundMutations.length === 1 ? ` ${mut} is` : `s ${mut} are`} mutated`,
-          stmt.loc,
-          {
-            extraDescription:
-              "Mutating function arguments has no effect outside the function due to call-by-value semantics",
-            suggestion:
-              "Return the modified value or use the contract's state to avoid unnecessary mutations",
-          },
-        ),
-      );
-    }
-    return acc;
+    return foundMutations.reduce((mutationMap, argName) => {
+      const existingStatements = mutationMap.get(argName) || [];
+      return mutationMap.set(argName, [...existingStatements, stmt]);
+    }, new Map<string, AstStatement[]>());
   }
 }